+++ lib/simplecov_mcp/error_handler.rb
@@ -12,7 +12,6 @@ module SimpleCovMcp
     VALID_ERROR_MODES = [:off, :log, :debug].freeze
 
     def initialize(error_mode: :log, logger: nil)
-      error_mode = normalize_mode(error_mode)
       unless VALID_ERROR_MODES.include?(error_mode)
         raise ArgumentError, "Invalid error_mode: #{error_mode.inspect}. Valid modes: #{VALID_ERROR_MODES.inspect}"
       end
@@ -161,17 +160,6 @@ module SimpleCovMcp
         "missing method '#{method_name}' on #{object_info}"
       else
         message
-      end
-    end
-
-    def normalize_mode(mode)
-      case mode
-      when :debug, 'debug', 'd', :trace, 'trace', 't'
-        :debug
-      when :log, 'log', 'l', :on, 'on'
-        :log
-      else
-        mode
       end
     end
   end